% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{art_write}
\alias{art_write}
\title{Save to file}
\usage{
art_write(input, filename, width = 10, height = 10, dpi = 300)
}
\arguments{
\item{input}{art object to print}

\item{width}{defaults to 3000 pixels}

\item{height}{defaults to 3000 pixels}

\item{file}{filename}

\item{background}{defaults to black}
}
\description{
Save to file
}
