% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridge}
\alias{bridge}
\title{Create animated brownian bridge}
\usage{
bridge(file, ntimes = 100, nseries = 20)
}
\arguments{
\item{file}{where to save}

\item{ntimes}{how many time points}

\item{nseries}{how many time series}
}
\description{
Create animated brownian bridge
}
