% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil.R, R/swirl.R
\name{coil}
\alias{coil}
\title{Coil}
\usage{
coil(bridge = meander(seed = 50, length = 50), scale = 0.02)

coil(bridge = meander(seed = 50, length = 50), scale = 0.02)
}
\arguments{
\item{bridge}{bridge tibble}

\item{scale}{length of swirl}

\item{bridge}{bridge tibble}

\item{scale}{length of swirl}
}
\description{
Coil

Coil
}
