% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collage.R
\name{collage}
\alias{collage}
\title{Creates a collage of noise samples}
\usage{
collage(generator = ambient::noise_worley, palette = "tokyo",
  grain = 500, ...)
}
\arguments{
\item{generator}{the function (in ambient) to generate noise}

\item{palette}{name of a scico palette}

\item{grain}{how detailed is each panel}

\item{...}{arguments to pass to the generator}
}
\value{
A ggplot object
}
\description{
Creates a collage of noise samples
}
