% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{export_animation}
\alias{export_animation}
\title{Export animation to a file}
\usage{
export_animation(input, filename, nframes = 200, detail = 5, type = "cairo")
}
\arguments{
\item{input}{art object to print}

\item{filename}{filename}

\item{nframes}{defaults to 200}

\item{detail}{number of interpolated frames, defaults to 5}

\item{type}{defaults to "cairo"}
}
\description{
Export animation to a file
}
