% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{export_image}
\alias{export_image}
\title{Export image to a file}
\usage{
export_image(input, filename, width = 10, height = 10, dpi = 300)
}
\arguments{
\item{input}{art object to print}

\item{filename}{filename}

\item{width}{defaults to 3000 pixels (10in at 300dpi)}

\item{height}{defaults to 3000 pixels (10in at 300dpi)}

\item{dpi}{defaults to 300dpi}
}
\description{
Export image to a file
}
