% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander.R
\name{flow_circle}
\alias{flow_circle}
\title{Rotate around the origin}
\usage{
flow_circle(theta = pi/128)
}
\arguments{
\item{theta}{size of rotation}
}
\value{
function that returns a modified data frame
}
\description{
Rotate around the origin
}
