% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{locate_entity}
\alias{locate_entity}
\title{Locate entities}
\usage{
locate_entity(entity, xpos = 0, ypos = 0, size = 1, angle = 0, ...)
}
\arguments{
\item{entity}{The entity to be placed}

\item{xpos}{The horizontal location of the entity}

\item{ypos}{The vertical location of the entity}

\item{size}{Parameter controlling the size of the entity}

\item{angle}{Parameter controlling the orientation of the entity}

\item{...}{Other arguments are ignored}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Locate entities
}
\details{
When a jasmine entity is created it is implicitly assumed to be
located at the origin (xpos = 0, ypos = 0), to have size 1, and to have
a horizontal orientation (angle = 0). The locate_entity function allows the
entity to be transformed in simple ways: translation, dilation and rotations
}
