% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_allvers}
\alias{make_allvers}
\title{Short cut for generating multiple versions of the file}
\usage{
make_allvers(file)
}
\arguments{
\item{file}{the original tiff file (8k x 8k)}
}
\description{
Short cut for generating multiple versions of the file
}
