% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander.R
\name{make_bridges}
\alias{make_bridges}
\title{Make smoothed brownian bridge time series}
\usage{
make_bridges(seed = 20, length = 100, smoothing = 6, endpause = 0)
}
\arguments{
\item{seed}{number of time series or complete seed as tibble}

\item{length}{number of time points in the time series}

\item{smoothing}{number of smoothing iterations}

\item{endpause}{length of pause at the end}
}
\value{
tibble with columns series, time, x, y
}
\description{
Make smoothed brownian bridge time series
}
