% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander.R
\name{make_bridges_from_seed}
\alias{make_bridges_from_seed}
\title{Smoothed brownian bridges with seed points}
\usage{
make_bridges_from_seed(dots, ntimes = 100, npause = 10)
}
\arguments{
\item{dots}{the seed points}

\item{ntimes}{times}

\item{npause}{length of pause at the end}
}
\value{
bridges tibble
}
\description{
Smoothed brownian bridges with seed points
}
