% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letters.R
\name{make_dotted_letters}
\alias{make_dotted_letters}
\title{Returns a tibble specifying letters as bitmaps}
\usage{
make_dotted_letters()
}
\value{
tibble
}
\description{
Returns a tibble specifying letters as bitmaps
}
