% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letters.R
\name{make_dotted_message}
\alias{make_dotted_message}
\title{Make a message in dotted text}
\usage{
make_dotted_message(text)
}
\arguments{
\item{text}{the message}
}
\value{
a tibble
}
\description{
Make a message in dotted text
}
