% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempest.R
\name{make_rows}
\alias{make_rows}
\title{Generate an even set of rows}
\usage{
make_rows(n = 10, grain = 1000)
}
\arguments{
\item{n}{how many rows}

\item{grain}{how many points along each row}
}
\value{
a tibble with columns x, y and id
}
\description{
Generate an even set of rows
}
