% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander.R
\name{meander}
\alias{meander}
\title{Create meandering walk animation}
\usage{
meander(file = NULL, bridge = make_bridges(), wake_length = 0.1,
  palette = palette_scico(palette = "berlin"), background = "black",
  ...)
}
\arguments{
\item{file}{where to save}

\item{bridge}{tibble specifying the time series}

\item{wake_length}{length of the tail}

\item{palette}{function generating palette values}

\item{background}{colour of the background}

\item{...}{other arguments to pass to shadow_wake}
}
\description{
Create meandering walk animation
}
