% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_manual}
\alias{palette_manual}
\title{Function factory for manual palettes}
\usage{
palette_manual(...)
}
\arguments{
\item{...}{colour names}
}
\value{
a function that takes arguments n and alpha
}
\description{
Function factory for manual palettes
}
