% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_named}
\alias{palette_named}
\title{Function factory for prespecified palettes}
\usage{
palette_named(name = NULL, ...)
}
\arguments{
\item{name}{name of the palette}

\item{...}{arguments to be passed to other functions}
}
\value{
a function that takes arguments n and alpha
}
\description{
Function factory for prespecified palettes
}
