% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_ropensci.R
\name{palette_ropensci}
\alias{palette_ropensci}
\title{Returns a paletting function}
\usage{
palette_ropensci(type = "plain")
}
\arguments{
\item{type}{which type (plain or whitened)}
}
\value{
a paletting function using ropensci colours
}
\description{
Returns a paletting function
}
