% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{palette_scico}
\alias{palette_scico}
\title{Function factory for scico palettes}
\usage{
palette_scico(...)
}
\arguments{
\item{...}{arguments (besides n and alpha) to be passed to scico::scico}
}
\value{
a function that takes arguments n and alpha
}
\description{
Function factory for scico palettes
}
