% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempest.R
\name{palette_viridis}
\alias{palette_viridis}
\title{Function factory viridis palettes}
\usage{
palette_viridis(...)
}
\arguments{
\item{...}{arguments to be passed to viridis::viridis}
}
\value{
a function that takes arguments n and alpha
}
\description{
Function factory viridis palettes
}
