% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_legacy.R
\name{scene_bubbles}
\alias{scene_bubbles}
\title{Seed with a random set of bubbles}
\usage{
scene_bubbles(n = 2, grain = 100)
}
\arguments{
\item{n}{Number of circles}

\item{grain}{The number of points per row}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Seed with a random set of bubbles
}
