% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_delaunay.R
\name{scene_delaunay}
\alias{scene_delaunay}
\title{Create a scene using Delaunay triangulation}
\usage{
scene_delaunay(seed = use_seed(1), n = 20, grain = 50)
}
\arguments{
\item{seed}{The RNG seed}

\item{n}{Number of vertices}

\item{grain}{Number of points along each line}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Create a scene using Delaunay triangulation
}
