% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes_legacy.R
\name{scene_discs}
\alias{scene_discs}
\title{Seed shaped like concentric rings}
\usage{
scene_discs(points = 100, rings = 3, size = 2)
}
\arguments{
\item{points}{Total number of interior points}

\item{rings}{How many rings to spread across?}

\item{size}{Diameter of the outermost ring}
}
\description{
Seed shaped like concentric rings
}
