% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes.R
\name{scene_grid}
\alias{scene_grid}
\title{Create a scene with multiple entities laid out in a grid}
\usage{
scene_grid(
  entity,
  nrow = 3,
  ncol = 3,
  grain = 50,
  size = 1,
  shape = 3,
  angle = 0
)
}
\arguments{
\item{entity}{The entity type (e.g., "line")}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{grain}{The number of points per entity}

\item{size}{The size of each entity}

\item{shape}{The shape of each entity}

\item{angle}{The angle of each entity}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Create a scene with multiple entities laid out in a grid
}
\details{
The \code{scene_grid()} function allows multiple entities to be
included in the initial object, laying out items in grid.
}
