% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_gridmix.R
\name{scene_mix}
\alias{scene_mix}
\title{Create a scene with entities placed randomly}
\usage{
scene_mix(
  seed = use_seed(1),
  n = 5,
  xpos = (1:20)/4,
  ypos = (1:20)/4,
  entity = c("circle", "line", "heart", "droplet"),
  grain = 100,
  size = (10:20)/20,
  shape = 3,
  angle = seq(0, 2 * pi, length.out = 20)
)
}
\arguments{
\item{seed}{Seed number to attach}

\item{n}{Number of entities}

\item{xpos}{Numeric vector specifying possible horizontal locations}

\item{ypos}{Numeric vector specifying possible vertical locations}

\item{entity}{Character vector specifying possible entity types (e.g., "line", "circle")}

\item{grain}{Numeric vector specifying possible grains}

\item{size}{Numeric vector specifying possible sizes}

\item{shape}{Numeric vector specifying possible shapes}

\item{angle}{Numeric vector specifying possible angles}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Create a scene with entities placed randomly
}
