% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes_legacy.R
\name{scene_rows}
\alias{scene_rows}
\title{Seed with evenly spaced rows}
\usage{
scene_rows(n = 10, grain = 100, vertical = FALSE)
}
\arguments{
\item{n}{Number of rows}

\item{grain}{The number of points per row}

\item{vertical}{Flip the x/y co-ords to produce columns?}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Seed with evenly spaced rows
}
