% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenes_legacy.R
\name{scene_sticks}
\alias{scene_sticks}
\title{Scene from a set of random sticks}
\usage{
scene_sticks(n = 10, grain = 100)
}
\arguments{
\item{n}{how many sticks}

\item{grain}{how many points along each stick}
}
\value{
a tibble with columns x, y and id
}
\description{
Scene from a set of random sticks
}
