% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_delaunay.R
\name{seed_delaunay}
\alias{seed_delaunay}
\title{Generate seed points with a Delaunay triangulation}
\usage{
seed_delaunay(n = 20, grain = 50)
}
\arguments{
\item{n}{Number of vertices}

\item{grain}{Number of points along each line}
}
\value{
A tibble with columns x, y and id
}
\description{
Generate seed points with a Delaunay triangulation
}
