% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_generators.R
\name{seed_disc}
\alias{seed_disc}
\title{Seed shaped like concentric rings}
\usage{
seed_disc(points = 1000, rings = 10, size = 1)
}
\arguments{
\item{points}{Total number of interior points}

\item{rings}{How many rings to spread across?}

\item{size}{Radius of the outermost ring}
}
\description{
Seed shaped like concentric rings
}
