% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_seeds.R
\name{seed_sticks}
\alias{seed_sticks}
\title{Generate a random set of sticks}
\usage{
seed_sticks(n = 10, grain = 1000)
}
\arguments{
\item{n}{how many sticks}

\item{grain}{how many points along each stick}
}
\value{
a tibble with columns x, y and id
}
\description{
Generate a random set of sticks
}
