% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_generators.R
\name{seed_teardrop}
\alias{seed_teardrop}
\title{Seed shaped like a teardrop}
\usage{
seed_teardrop(n = 100, m = 3)
}
\arguments{
\item{n}{number of dot points}

\item{m}{exponent}
}
\value{
tibble
}
\description{
Seed shaped like a teardrop
}
