% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline.R
\name{skyline}
\alias{skyline}
\title{Animated GIF of Sydney skyline}
\usage{
skyline(file)
}
\arguments{
\item{file}{path to gif}
}
\description{
Animated GIF of Sydney skyline
}
\details{
Creates an animated gif of the Sydney skyline
}
