% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_overlay.R
\name{style_overlay}
\alias{style_overlay}
\title{Adds an overlay to the existing plot}
\usage{
style_overlay(pic, border = NULL, fill = NULL, linewidth = 1)
}
\arguments{
\item{pic}{Existing plot}

\item{border}{Colour of border}

\item{fill}{Colour of fill}

\item{linewidth}{Width of border}
}
\description{
Adds an overlay to the existing plot
}
