% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_pop.R
\name{style_pop}
\alias{style_pop}
\title{Style as a pop art image}
\usage{
style_pop(
  data,
  palette = "base",
  colour = "ind",
  alpha = 0.3,
  fade = 0,
  background = "warhol",
  adjust = 0.7,
  panels = 4,
  ...
)
}
\arguments{
\item{data}{data frame with x, y, order, id, time}

\item{palette}{function generating a palette (or string naming the palette)}

\item{colour}{name of variable to use to specify the colour aesthetic}

\item{alpha}{length two numeric, first element is the initial alpha, (optional) second is the decay rate for alpha}

\item{background}{colour of the background in the plot}

\item{...}{arguments to pass to geom}
}
\value{
Returns a ggplot2 object
}
\description{
Style as a pop art image
}
