% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_wrap.R
\name{style_wrap}
\alias{style_wrap}
\title{Style as a wrap image}
\usage{
style_wrap(data, background = "black", type = "line",
  palette = palette_viridis(), size = 2)
}
\arguments{
\item{data}{a coiled bridge}

\item{background}{colour}

\item{type}{line, barbelle or points}

\item{palette}{palette generator}

\item{size}{line width}
}
\value{
returns a ggplot2 object
}
\description{
Style as a wrap image
}
