% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swirl.R
\name{swirl}
\alias{swirl}
\title{Swirl}
\usage{
swirl(bridge = make_bridges(50, 50), scale = 0.02, type = "line")
}
\arguments{
\item{bridge}{bridge tibble}

\item{scale}{length of swirl}

\item{type}{plot type (line, point, barbell)}
}
\description{
Swirl
}
