% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempest.R
\name{tempest}
\alias{tempest}
\title{Generate a tempest image}
\usage{
tempest(file = NULL, seed = make_sticks(), iterations = 6,
  burnin = 0, scale = 0.02, alpha_init = 0.3, alpha_decay = 0,
  width = 6, box = NULL, zoom = TRUE, seed_col = NULL,
  seed_fill = NULL, palette = NULL, ...)
}
\arguments{
\item{file}{where to save the file (default: NULL)}

\item{seed}{data frame with x, y, id}

\item{iterations}{how many times should we iterate the curl noise?}

\item{burnin}{how many iterations should we discard as burnin?}

\item{scale}{how large is each curl step?}

\item{alpha_init}{how transparent is each line?}

\item{alpha_decay}{how does transparency decay over iterations?}

\item{width}{how wide are the lines?}

\item{box}{vector specifying the edges}

\item{zoom}{rescale image to fit in the unit square?}

\item{seed_col}{what colour to draw the seed? (default: NULL)}

\item{seed_fill}{what colour to fill the seed? (default: NULL)}

\item{palette}{function generating a palette (default: viridis)}

\item{...}{arguments to be passed to palette}
}
\description{
Generate a tempest image
}
