% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_reverse.R
\name{time_reverse}
\alias{time_reverse}
\title{Transform a temporal object by time reversing it}
\usage{
time_reverse(data, append = FALSE)
}
\arguments{
\item{data}{the object}

\item{append}{should the time reversed version be appended (loop back) or substituted?}
}
\value{
The same object but with time indices reversed
}
\description{
Transform a temporal object by time reversing it
}
