% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_breeze.R
\name{unfold_breeze}
\alias{unfold_breeze}
\title{Unfold a scene with a swirl operation}
\usage{
unfold_breeze(
  data = scene_sticks(),
  iterations = 6,
  scale = 0.02,
  drift = 0.01,
  noise = NULL,
  fractal = NULL,
  octaves = 8,
  output = "time",
  ...
)
}
\arguments{
\item{data}{data frame with x, y, id, and more}

\item{iterations}{how many times should we iterate?}

\item{scale}{how large is each step?}

\item{drift}{gaussian noise to inject at each step}

\item{noise}{noise function (default is ambient::gen_simplex)}

\item{fractal}{fractal function (default is ambient::billow)}

\item{octaves}{default = 8}

\item{output}{name of the unfolding variable to add (e.g., time)}

\item{...}{arguments to pass to ambient::fracture}
}
\value{
a "tempest" ribbon, data frame with x, y, order, time and id
}
\description{
Unfold a scene with a swirl operation
}
