% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_loop.R
\name{unfold_loop}
\alias{unfold_loop}
\title{Take a seed object and convert each point to a loop}
\usage{
unfold_loop(data, points = 20, radius = 1)
}
\arguments{
\item{data}{data}

\item{points}{number of time points}

\item{radius}{radius of the circle}
}
\value{
a tibble with x, y, id and time
}
\description{
Take a seed object and convert each point to a loop
}
