% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_meander.R
\name{unfold_meander}
\alias{unfold_meander}
\title{Make smoothed brownian bridge time series}
\usage{
unfold_meander(
  data = 20,
  iterations = 100,
  smoothing = 6,
  endpause = 0,
  output1 = "time",
  output2 = "series"
)
}
\arguments{
\item{data}{data frame with x, y, etc}

\item{iterations}{number of time points in the time series}

\item{smoothing}{number of smoothing iterations}

\item{endpause}{length of pause at the end}

\item{output}{name of the variable}
}
\value{
tibble with columns series, time, x, y
}
\description{
Make smoothed brownian bridge time series
}
