% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_slice.R
\name{unfold_slice}
\alias{unfold_slice}
\title{Unfold a scene with a slice operation}
\usage{
unfold_slice(
  data = scene_sticks(),
  iterations = 6,
  scale = 0.2,
  scatter = FALSE,
  output1 = "time",
  output2 = "order"
)
}
\arguments{
\item{data}{data frame with x, y, id, and more}

\item{iterations}{how many times should we iterate the curl noise?}

\item{scale}{how large is each curl step?}

\item{scatter}{should the noise seed be "scattered"?}

\item{output1}{name of the primary unfolding variable to add (e.g., time)}

\item{output2}{name of the secondary unfolding variable to add (e.g., order)}
}
\value{
a "tempest" ribbon, data frame with x, y, order, time and id
}
\description{
Unfold a scene with a slice operation
}
