% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_tempest.R
\name{unfold_tempest}
\alias{unfold_tempest}
\title{Generate a tempest object}
\usage{
unfold_tempest(
  data = seed_sticks(),
  iterations = 6,
  scale = 0.02,
  seed = NULL,
  output1 = "time",
  output2 = "order"
)
}
\arguments{
\item{data}{data frame with x, y, id, and more}

\item{iterations}{how many times should we iterate the curl noise?}

\item{scale}{how large is each curl step?}

\item{seed}{seed to pass to curl_noise()}
}
\value{
a "tempest" ribbon, data frame with x, y, order, time and id
}
\description{
Generate a tempest object
}
