% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_what.R
\name{unfold_what}
\alias{unfold_what}
\title{Generate WHAT?}
\usage{
unfold_what(
  data = scene_sticks(),
  iterations = 6,
  scale = 0.02,
  scatter = FALSE,
  output = "time"
)
}
\arguments{
\item{data}{data frame with x, y, id, and more}

\item{iterations}{how many times should we iterate the noise?}

\item{scale}{how large is each step?}

\item{scatter}{should the noise seed be "scattered"?}

\item{output}{name of the output variable (default = time)}
}
\value{
thing
}
\description{
Generate WHAT?
}
