% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_worley.R
\name{unfold_worley}
\alias{unfold_worley}
\title{Colour points using Worley noise}
\usage{
unfold_worley(data, seed = NULL, output = "order", ...)
}
\arguments{
\item{data}{the data object}

\item{seed}{seed to pass to ambient::gen_worley()}

\item{output}{name of the primary unfolding variable to add (e.g., order)}

\item{...}{arguments to pass to ambient::genworley()}
}
\description{
Colour points using Worley noise
}
