% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_seed}
\alias{use_seed}
\title{Wrapper to set.seed}
\usage{
use_seed(seed, ...)
}
\arguments{
\item{seed}{The seed number}

\item{...}{Other arguments to set.seed()}
}
\description{
Wrapper to set.seed
}
\details{
Calling use_seed has two side effects: it calls set.seed to set the
seed for the RNG, and it also records the value of "seed" in as the "jasmine.seed"
option. Once specified, any function in jasmines with a seed argument will use this
value unless manually set by the user.
}
