% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{art_write}
\alias{art_write}
\title{Save to file}
\usage{
art_write(input, filename, width = 10, height = 10, dpi = 300)
}
\arguments{
\item{input}{art object to print}

\item{filename}{filename}

\item{width}{defaults to 3000 pixels (10in at 300dpi)}

\item{height}{defaults to 3000 pixels (10in at 300dpi)}

\item{dpi}{defaults to 300dpi}
}
\description{
Save to file
}
