% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hex.R
\name{as_hex}
\alias{as_hex}
\title{Convert image to hex}
\usage{
as_hex(
  from,
  to,
  background = "#ffffff00",
  crop = 1,
  text = NULL,
  text_colour = "white",
  text_size = 50
)
}
\arguments{
\item{from}{Path to original file}

\item{to}{Path to destination file}

\item{background}{Background colour}

\item{crop}{Proportion of (smaller dimension of) image to crop to}

\item{text}{Text annotation}

\item{text_colour}{Colour of text annotation}
}
\description{
Convert image to hex
}
