% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hex.R
\name{as_hex}
\alias{as_hex}
\title{Convert image to hex}
\usage{
as_hex(
  from,
  to,
  text_label = NULL,
  text_colour = "white",
  border_colour = "grey20",
  border_opacity = 60
)
}
\arguments{
\item{from}{Path to original file}

\item{to}{Path to destination file}

\item{text_label}{Text annotation}

\item{text_colour}{Colour of text annotation}

\item{border_colour}{Colour of the border region}

\item{border_opacity}{Opacity of the border region}
}
\description{
Convert image to hex
}
