% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{entitytype}
\alias{entitytype}
\alias{entity_circle}
\alias{entity_line}
\alias{entity_heart}
\alias{entity_droplet}
\alias{entity_gaussian}
\alias{entity_null}
\title{Entity types}
\usage{
entity_circle(grain = 50, id = NULL, ...)

entity_line(grain = 50, id = NULL, ...)

entity_heart(grain = 50, id = NULL, ...)

entity_droplet(grain = 50, id = NULL, shape = 3, ...)

entity_gaussian(grain = 50, id = NULL, ...)

entity_null(...)
}
\arguments{
\item{grain}{The number of points that comprise the entity}

\item{id}{A numeric identifier for the entity}

\item{...}{Parameters to be passed to locate_entity}

\item{shape}{Parameter controlling the shape of the entity (droplet only)}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Entity types
}
\details{
Primitive entities in jasmines are tibbles with four columns: x and y specify
co-ordinate values, the id is a number identifying the object, and the type is
a character label indicating what kind of entity it is. By default, entities are
assigned a random integer as the id code, but it is often wise for the calling
function to assign the id in a more predictable fashion.
}
