% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{export_image}
\alias{export_image}
\title{Export image to a file}
\usage{
export_image(
  input,
  filename,
  width = 10,
  height = 10,
  dpi = 300,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{input}{art object to print}

\item{filename}{filename}

\item{width}{defaults to 3000 pixels (10in at 300dpi)}

\item{height}{defaults to 3000 pixels (10in at 300dpi)}

\item{dpi}{defaults to 300dpi}

\item{xlim}{by default plot limits are c(-.05, 1.05), relative to data spanning c(0, 1), but can override}

\item{ylim}{by default plot limits are c(-.05, 1.05), relative to data spanning c(0, 1), but can override}
}
\description{
Export image to a file
}
