% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meander.R
\name{make_bridges}
\alias{make_bridges}
\title{Make smoothed brownian bridge time series}
\usage{
make_bridges(ntimes = 100, nseries = 20, smoothing = 6)
}
\arguments{
\item{ntimes}{number of time points}

\item{nseries}{number of time series}

\item{smoothing}{number of smoothing iterations}
}
\value{
tibble with columns series, time, x, y
}
\description{
Make smoothed brownian bridge time series
}
