% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_adjust}
\alias{palette_adjust}
\title{Function factory for adjusted palettes}
\usage{
palette_adjust(name, prefix, ...)
}
\arguments{
\item{name}{Name of the base palette}

\item{prefix}{Vector of colours to preprend to the named palette}

\item{...}{Arguments to be passed to grDevices::adjustcolor}
}
\value{
A modified
}
\description{
Function factory for adjusted palettes
}
