% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_legacy.R
\name{scene_bubbles}
\alias{scene_bubbles}
\title{Create a scene comprised of circles of varying size and location}
\usage{
scene_bubbles(seed = use_seed(1), n = 2, grain = 100)
}
\arguments{
\item{seed}{Seed number to attach}

\item{n}{Number of circles}

\item{grain}{The number of points per row}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Create a scene comprised of circles of varying size and location
}
