% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_legacy.R
\name{scene_discs}
\alias{scene_discs}
\title{Create a scene comprised of concentric circles}
\usage{
scene_discs(seed = use_seed(1), points = 100, rings = 3, size = 2)
}
\arguments{
\item{seed}{Seed number to attach}

\item{points}{Total number of interior points}

\item{rings}{How many rings to spread across?}

\item{size}{Diameter of the outermost ring}
}
\description{
Create a scene comprised of concentric circles
}
