% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_gridmix.R
\name{scene_grid}
\alias{scene_grid}
\title{Create a scene with entities on a grid}
\usage{
scene_grid(
  seed = use_seed(1),
  xpos = 1:3,
  ypos = 1:3,
  entity = "circle",
  grain = 50,
  size = 1,
  shape = 3,
  angle = 0
)
}
\arguments{
\item{seed}{Seed number to attach}

\item{xpos}{Numeric vector specifying horizontal locations}

\item{ypos}{Numeric vector specifying vertical locations}

\item{entity}{The entity type (e.g., "line", "circle")}

\item{grain}{The number of points per entity}

\item{size}{The size of each entity}

\item{shape}{The shape of each entity}

\item{angle}{The angle of each entity}
}
\value{
A tibble with four columns: x, y, id and type
}
\description{
Create a scene with entities on a grid
}
\details{
The \code{scene_grid()} function allows multiple entities to be
included in the initial object, laying out items in grid.
}
