% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_legacy.R
\name{scene_rows}
\alias{scene_rows}
\title{Create a scene comprised of horizontal or vertical lines}
\usage{
scene_rows(seed = use_seed(1), n = 10, grain = 100, vertical = FALSE)
}
\arguments{
\item{seed}{Seed number to attach}

\item{n}{Number of rows}

\item{grain}{The number of points per row}

\item{vertical}{Flip the x/y co-ords to produce columns?}
}
\value{
A tibble with columns: x, y, id, type, seed
}
\description{
Create a scene comprised of horizontal or vertical lines
}
