% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_legacy.R
\name{scene_sticks}
\alias{scene_sticks}
\title{Create a scene comprised of lines of varying length and orientation}
\usage{
scene_sticks(seed = use_seed(1), n = 10, grain = 100)
}
\arguments{
\item{seed}{Seed number to attach}

\item{n}{how many sticks}

\item{grain}{how many points along each stick}
}
\value{
a tibble with columns x, y and id
}
\description{
Create a scene comprised of lines of varying length and orientation
}
