% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_seeds.R
\name{seed_bubbles}
\alias{seed_bubbles}
\title{Generate a random set of bubbles}
\usage{
seed_bubbles(n = 2, grain = 1000)
}
\arguments{
\item{n}{how many bubbles}

\item{grain}{how many points along each stick}
}
\value{
a tibble with columns x, y and id
}
\description{
Generate a random set of bubbles
}
