% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{seed_rows}
\alias{seed_rows}
\title{Seed with evenly spaced rows}
\usage{
seed_rows(n = 10, grain = 1000)
}
\arguments{
\item{n}{how many rows}

\item{grain}{how many points along each row}
}
\value{
a tibble with columns x, y and id
}
\description{
Seed with evenly spaced rows
}
