% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_generators.R
\name{seed_sticks}
\alias{seed_sticks}
\title{Seed from a set of random sticks}
\usage{
seed_sticks(n = 10, grain = 1000)
}
\arguments{
\item{n}{how many sticks}

\item{grain}{how many points along each stick}
}
\value{
a tibble with columns x, y and id
}
\description{
Seed from a set of random sticks
}
