% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_ribbon.R
\name{style_ribbon}
\alias{style_ribbon}
\title{Style as a ribbon image}
\usage{
style_ribbon(
  data,
  burnin = 0,
  alpha_init = 0.3,
  alpha_decay = 0,
  seed_col = NULL,
  seed_fill = NULL,
  seed_linewidth = 1,
  palette = palette_viridis(),
  background = "black",
  type = "segment",
  ...
)
}
\arguments{
\item{data}{data frame with x, y, order, id, time}

\item{burnin}{how many iterations should we discard as burnin?}

\item{alpha_init}{how transparent is each line?}

\item{alpha_decay}{how does transparency decay over iterations?}

\item{seed_col}{what colour to draw the seed? (default: NULL)}

\item{seed_fill}{what colour to fill the seed? (default: NULL)}

\item{seed_linewidth}{width of the seed lines}

\item{palette}{function generating a palette}

\item{background}{colour of the background in the plot}

\item{type}{type of geom to use ("segment", "curve" or "point")}

\item{...}{arguments to pass to geom}
}
\value{
Returns a ggplot2 object
}
\description{
Style as a ribbon image
}
