% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_ribbon.R, R/tempest.R
\name{style_ribbon}
\alias{style_ribbon}
\title{Draw a ribbon image}
\usage{
style_ribbon(ribbon, file = NULL, burnin = 0, alpha_init = 0.3,
  alpha_decay = 0, line_width = 6, box = NULL, zoom = TRUE,
  seed_col = NULL, seed_fill = NULL, palette = palette_viridis(),
  width = 3000, height = 3000, background = "black")

style_ribbon(ribbon, file = NULL, burnin = 0, alpha_init = 0.3,
  alpha_decay = 0, line_width = 6, box = NULL, zoom = TRUE,
  seed_col = NULL, seed_fill = NULL, palette = palette_viridis(),
  width = 3000, height = 3000, background = "black")
}
\arguments{
\item{ribbon}{data frame with x, y, order, id, time}

\item{file}{path to tiff output file or NULL (the default) to display on screen}

\item{burnin}{how many iterations should we discard as burnin?}

\item{alpha_init}{how transparent is each line?}

\item{alpha_decay}{how does transparency decay over iterations?}

\item{line_width}{how wide are the lines?}

\item{box}{vector specifying the edges}

\item{zoom}{rescale image to fit in the unit square?}

\item{seed_col}{what colour to draw the seed? (default: NULL)}

\item{seed_fill}{what colour to fill the seed? (default: NULL)}

\item{palette}{function generating a palette, or a list of such functions, with length equal to the number of objects in the seed}

\item{width}{width in pixels (default = 3000, for printing try 8000)}

\item{height}{height in pixels (default = 3000, for printing try 8000)}

\item{background}{colour of the background in the plot}

\item{ribbon}{data frame with x, y, order, id, time}

\item{file}{path to tiff output file or NULL (the default) to display on screen}

\item{burnin}{how many iterations should we discard as burnin?}

\item{alpha_init}{how transparent is each line?}

\item{alpha_decay}{how does transparency decay over iterations?}

\item{line_width}{how wide are the lines?}

\item{seed_col}{what colour to draw the seed? (default: NULL)}

\item{seed_fill}{what colour to fill the seed? (default: NULL)}

\item{palette}{function generating a palette, or a list of such functions, with length equal to the number of objects in the seed}

\item{width}{width in pixels (default = 3000, for printing try 8000)}

\item{height}{height in pixels (default = 3000, for printing try 8000)}

\item{background}{colour of the background in the plot}
}
\description{
Draw a ribbon image

Draw a ribbon image
}
