% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_wrap.R, R/swirl.R
\name{style_wrap}
\alias{style_wrap}
\title{Style wrap}
\usage{
style_wrap(coil, file = NULL, background = "black", type = "line",
  width = 3000, height = 3000, palette = palette_viridis())

style_wrap(coil, file = NULL, background = "black", type = "line",
  width = 3000, height = 3000, palette = palette_viridis())
}
\arguments{
\item{coil}{a coiled bridge}

\item{background}{colour}

\item{type}{line, barbelle or points}

\item{palette}{palette generator}

\item{coil}{a coiled bridge}

\item{background}{colour}

\item{type}{line, barbelle or points}

\item{palette}{palette generator}
}
\value{
ggplot

ggplot
}
\description{
Style wrap

Style wrap
}
