% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_mono}
\alias{theme_mono}
\title{Theme that supplies a background colour only}
\usage{
theme_mono(background)
}
\arguments{
\item{background}{the background colour}
}
\value{
ggplot theme
}
\description{
Theme that supplies a background colour only
}
