% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_mono}
\alias{theme_mono}
\title{Monocolour version of theme_void}
\usage{
theme_mono(background)
}
\arguments{
\item{background}{}
}
\value{
ggplot theme
}
\description{
Monocolour version of theme_void
}
