% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_inside.R
\name{unfold_inside}
\alias{unfold_inside}
\title{Unfold a scene with an inside operation}
\usage{
unfold_inside(data, output = "inside")
}
\arguments{
\item{data}{Data}

\item{output}{String specifying a Column name}
}
\value{
Returns the original data tibble with a new column added
}
\description{
Unfold a scene with an inside operation
}
