% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_meander.R
\name{unfold_meander}
\alias{unfold_meander}
\title{Unfold a scene with a meander operation}
\usage{
unfold_meander(
  data = entity_circle(),
  iterations = 100,
  smoothing = 6,
  endpause = 0,
  output1 = "time",
  output2 = "series"
)
}
\arguments{
\item{data}{data frame with x, y, etc}

\item{iterations}{number of time points in the time series}

\item{smoothing}{number of smoothing iterations}

\item{endpause}{length of pause at the end}

\item{output1}{name of the primary unfolding variable to add (e.g., time)}

\item{output2}{name of the secondary unfolding variable to add (e.g., series)}
}
\value{
tibble with columns series, time, x, y
}
\description{
Unfold a scene with a meander operation
}
