% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold_worley.R
\name{unfold_worley}
\alias{unfold_worley}
\title{Unfold a scene with a worley operation}
\usage{
unfold_worley(data, scatter = FALSE, output = "order", ...)
}
\arguments{
\item{data}{the data object}

\item{scatter}{should the noise seed be "scattered"?}

\item{output}{name of the primary unfolding variable to add (e.g., order)}

\item{...}{arguments to pass to ambient::genworley()}
}
\description{
Unfold a scene with a worley operation
}
